// src/app/shared/swagger/model/editTlInqueryDto.ts
/**
 * Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EditTlInqueryDto { 
    id: number;
    sped: string;
    country: string;
    acceptingPort: string;
    expectedRetrieveWeek: string;
    invoiceOn: string;
    retrieveDate: string;
    retrieveLocation: string;
    scGeneral: number;
    scMain: number;
    scTrail: number;
    portOfDeparture: string;
    ets: string;
    eta: string;
    boat: string;
}